extract multi devices graph builder
